requirements.txt
fastapi
requests
jinja2
python-dotenv
uvicorn
